Separate model files.
