+++ lib/oxygen_auth.js
@@ -1,5 +1,5 @@
 var passport = require('passport')
-  , UserModel = require('./models').UserModel
+  , UserModel = require('./models/user').UserModel
   , util = require('util')
   , request = require('superagent');
