+++ lib/packages.js
@@ -1,10 +1,10 @@
-var PackageModel = require('./models').PackageModel
+var PackageModel = require('./models/package').PackageModel
 	, error = require('./error')
 	, users = require('./users')
 	, async = require('async')
 	, _ = require('underscore')
 	, mongoose = require('mongoose')
-	, UserModel = require('../lib/models').UserModel
+	, UserModel = require('../models/user').UserModel
     , fs = require('fs')
 	, crypto = require('crypto')
     , storage = require('./storage');
@@ -1687,7 +1687,7 @@ function _delete_pkg(pkg_id, callback) {
  * @param {Function} Callback to execute after white listing. The arguments are an error object (null if successful) and the white listed pkg.
  * @api private
  */
-exports.whitelist_by_id = function(pkg_id, callback){
+exports.whitelist_by_id = function(pkg_id, add, callback){
 	PackageModel.findById(pkg_id, function(err,pkg){
 		if(err){
 			if(callback){
@@ -1695,47 +1695,17 @@ exports.whitelist_by_id = function(pkg_id, callback){
 				return;
 			}
 		}
-
-	    pkg.white_listed = true;
+        
+	    pkg.white_listed = add;
+        
 	    pkg.save(function(err){
 	      if(err){
 	      	if(callback){
-	      		callback(error.fail('There was an error white-listing the package.'));
+	      		callback(error.fail('There was an error setting the white-list value for package.'));
 	      		return;
 	      	}
 	      }
 	    });
-
-		if(callback){
-			callback(null, pkg);
-		}
-	});
-};
-
-/**
- * Remove a package from the white list
- *
- * @param {ObjectId} The id of the pkg
- * @param {Function} Callback to execute after removal from the white list. The arguments are an error object (null if successful) and the pkg removed from the white list.
- * @api private
- */
-exports.unwhitelist_by_id = function(pkg_id, callback){
-	PackageModel.findById(pkg_id, function(err,pkg){
-		if(err){
-			if(callback){
-				callback(error.fail('Failed to find the package in the db.'))
-				return;
-			}
-		}
-
-	    pkg.white_listed = false;
-    	pkg.save(function(err){
-	      if(err){
-	      	if(callback){
-	      		callback(error.fail('There was an error white-listing the package.'));
-	      	}
-	      }
-    	});
 
 		if(callback){
 			callback(null, pkg);
