+++ lib/stats_update.js
@@ -1,5 +1,5 @@
-var PackageModel = require('../lib/models').PackageModel
-	, UserModel = require('../lib/models').UserModel
+var PackageModel = require('../models/package').PackageModel
+	, UserModel = require('../models/user').UserModel
   , error = require('./error')
   , packages = require('./packages')
   , mongoose = require('mongoose')
