+++ lib/storage.js
@@ -3,7 +3,7 @@ var error = require('./error')
 	, async = require('async')
 	, _ = require('underscore')
 	, amazonS3 = require('awssum-amazon-s3')
-	, UserModel = require('../lib/models').UserModel
+	, UserModel = require('../models/user').UserModel
     , fs = require('fs.extra')
     , fmt = require('fmt')
 	, crypto = require('crypto');
