+++ lib/users.js
@@ -1,5 +1,5 @@
-var UserModel = require('./models').UserModel
-	, PackageModel = require('./models').PackageModel
+var UserModel = require('./models/user').UserModel
+	, PackageModel = require('./models/package').PackageModel
 	, error = require('./error')
 	, _ = require('underscore')
 	, async = require('async');
